split currency management in dedicated classes
