+++ bootstrap/src/main/java/onlydust/com/marketplace/api/bootstrap/configuration/BackofficeConfiguration.java
@@ -4,6 +4,7 @@ import onlydust.com.marketplace.accounting.domain.port.in.CurrencyFacadePort;
 import onlydust.com.marketplace.api.domain.port.input.BackofficeFacadePort;
 import onlydust.com.marketplace.api.domain.port.output.BackofficeStoragePort;
 import onlydust.com.marketplace.api.domain.service.BackofficeService;
+import onlydust.com.marketplace.api.rest.api.adapter.BackofficeCurrencyManagementRestApi;
 import onlydust.com.marketplace.api.rest.api.adapter.BackofficeRestApi;
 import onlydust.com.marketplace.api.rest.api.adapter.authentication.api_key.ApiKeyAuthenticationService;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -16,9 +17,13 @@ import org.springframework.context.annotation.Profile;
 public class BackofficeConfiguration {
 
     @Bean
-    public BackofficeRestApi backofficeRestApi(final BackofficeFacadePort backofficeFacadePort,
-                                               final CurrencyFacadePort currencyFacadePort) {
-        return new BackofficeRestApi(backofficeFacadePort, currencyFacadePort);
+    public BackofficeRestApi backofficeRestApi(final BackofficeFacadePort backofficeFacadePort) {
+        return new BackofficeRestApi(backofficeFacadePort);
+    }
+
+    @Bean
+    public BackofficeCurrencyManagementRestApi backofficeCurrencyManagementRestApi(final CurrencyFacadePort currencyFacadePort) {
+        return new BackofficeCurrencyManagementRestApi(currencyFacadePort);
     }
 
     @Bean
